데이터타입수정

+++ src/main/java/egovframework/com/global/file/util/FileUtils.java
@@ -127,7 +127,7 @@ public class FileUtils {
 		    throw new FileNotFoundException(downFileName);
 		}
 		
-		int fileSize = ((BigDecimal)request.getAttribute("fileSize")).intValue();
+		int fileSize = ((int)request.getAttribute("fileSize"));
 	
 		byte[] b = new byte[BUFF_SIZE]; //buffer size 2K.
